+++ openeo/rest/datacube.py
@@ -809,7 +809,7 @@ class DataCube(_ProcessGraphAbstraction):
     @openeo_process
     def resample_spatial(
         self,
-        resolution: Union[float, Tuple[float, float], List[float]] = 0.0,
+        resolution: Union[float, Tuple[float, float], List[float], Parameter] = 0.0,
         projection: Union[int, str, None] = None,
         method: str = "near",
         align: str = "upper-left",