+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
-
+- `DataCube.resample_spatial()` now supports `openeo.api.process.Parameter` objects for the `resolution` parameter.  ([#897](https://github.com/Open-EO/openeo-python-client/issues/897))
 
 ## [0.49.0] - 2026-04-01